deleted useless line of codes , added file utils.ts , renamed file's name , optimized code
