+++ src/Components/Drawer/types.ts
@@ -1,0 +1,3 @@
+export type TDrawer = {
+  setMobileOpen: (isOpen: boolean) => void;
+};
