+++ src/consts/sidebarCrm.tsx
@@ -3,7 +3,6 @@ import CleaningServicesIcon from "@mui/icons-material/CleaningServices";
 import FreeBreakfastIcon from "@mui/icons-material/FreeBreakfast";
 import CakeIcon from "@mui/icons-material/Cake";
 import PieChartIcon from "@mui/icons-material/PieChart";
-
 import { EUrls } from "./urls";
 import { TCrmList } from "./types";
